feat(dashboards): move paid media sections to the paid media drawer
The Attribution drawer rendered three paid-media sections — Paid Media by
Project, Attribution Channels (impression share by paid channel, despite
the name) and Paid Media Trend. All three describe paid delivery, not
attribution, so they belong with the Paid Media card.

The Paid Media drawer now takes the revenueImpact payload as an input,
wired from the overview alongside the Attribution drawer, and owns the
chart options, project-breakdown computeds and channel-label helpers those
sections need. The Attribution drawer keeps Marketing Attribution, Revenue
by Channel and Event Registration Attribution, and drops the members left
orphaned by the move.

LFXV2-2023
