+++ apps/lfx-one/src/app/modules/dashboards/executive-director/components/paid-social-reach-drawer/paid-social-reach-drawer.component.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 import { DecimalPipe } from '@angular/common';
-import { ChangeDetectionStrategy, Component, computed, inject, model, signal, Signal } from '@angular/core';
+import { ChangeDetectionStrategy, Component, computed, inject, input, model, signal, Signal } from '@angular/core';
 import { toObservable, toSignal } from '@angular/core/rxjs-interop';
 import { ButtonComponent } from '@components/button/button.component';
 import { CardComponent } from '@components/card/card.component';
@@ -18,7 +18,15 @@ import { DrawerModule } from 'primeng/drawer';
 import { SkeletonModule } from 'primeng/skeleton';
 
 import type { ChartData, ChartOptions } from 'chart.js';
-import type { SocialReachResponse, MarketingRecommendedAction, MarketingKeyInsight } from '@lfx-one/shared/interfaces';
+import type {
+  MarketingKeyInsight,
+  MarketingRecommendedAction,
+  RevenueImpactAttributionChannelView,
+  RevenueImpactChannelLegendView,
+  RevenueImpactProjectBreakdownView,
+  RevenueImpactResponse,
+  SocialReachResponse,
+} from '@lfx-one/shared/interfaces';
 
 @Component({
   selector: 'lfx-paid-social-reach-drawer',
@@ -35,6 +43,79 @@ export class PaidSocialReachDrawerComponent {
 
   // === Model Signals (two-way binding) ===
   public readonly visible = model<boolean>(false);
+
+  // === Inputs ===
+  /** Revenue-impact payload — carries the paid-media project breakdown, channel mix
+   *  and spend/revenue trend that this drawer renders below its own social-reach data. */
+  public readonly data = input<RevenueImpactResponse>({
+    pipelineInfluenced: 0,
+    revenueAttributed: 0,
+    matchRate: 0,
+    changePercentage: 0,
+    trend: 'up',
+    attributionModels: { linear: 0, firstTouch: 0, lastTouch: 0 },
+    engagementTypes: [],
+    paidMedia: { roas: 0, impressions: 0, adSpend: 0, adRevenue: 0, monthlyTrend: [] },
+    attributionChannels: [],
+    projectBreakdown: [],
+    eventRegistrationAttribution: { channelBreakdown: [], monthlyTrend: [] },
+  });
+
+  private static readonly channelBgClass: Record<string, string> = {
+    google_ads: 'bg-blue-500',
+    facebook_ads: 'bg-blue-700',
+    microsoft_ads: 'bg-emerald-600',
+    linkedin_ads: 'bg-gray-700',
+    reddit_ads: 'bg-red-500',
+    twitter_ads: 'bg-gray-500',
+  };
+  private static readonly channelBgFallback = 'bg-gray-400';
+
+  protected readonly paidMediaTrendChartOptions: ChartOptions<'bar'> = {
+    responsive: true,
+    maintainAspectRatio: false,
+    interaction: { mode: 'index', intersect: false },
+    plugins: {
+      legend: {
+        display: true,
+        position: 'top',
+        align: 'end',
+        labels: { color: lfxColors.gray[700], font: { size: 11 }, boxWidth: 12, boxHeight: 12, padding: 12 },
+      },
+      tooltip: {
+        ...DASHBOARD_TOOLTIP_CONFIG,
+        callbacks: {
+          label: (ctx) => ` ${ctx.dataset.label ?? ''}: $${Number(ctx.parsed.y ?? 0).toLocaleString()}`,
+        },
+      },
+    },
+    scales: {
+      x: {
+        display: true,
+        grid: { display: false },
+        border: { display: true, color: lfxColors.gray[300] },
+        ticks: { color: lfxColors.gray[500], font: { size: 11 } },
+      },
+      y: {
+        type: 'linear',
+        position: 'left',
+        display: true,
+        grid: { color: lfxColors.gray[200], lineWidth: 1 },
+        border: { display: false },
+        title: { display: true, text: 'Dollars ($)', color: lfxColors.gray[500], font: { size: 11 } },
+        ticks: {
+          color: lfxColors.gray[500],
+          font: { size: 11 },
+          callback: (value) => {
+            const n = Number(value);
+            if (n >= 1_000_000) return `$${(n / 1_000_000).toFixed(1)}M`;
+            if (n >= 1_000) return `$${(n / 1_000).toFixed(0)}K`;
+            return `$${n.toLocaleString()}`;
+          },
+        },
+      },
+    },
+  };
 
   // === WritableSignals ===
   protected readonly drawerLoading = signal(false);
@@ -163,6 +244,18 @@ export class PaidSocialReachDrawerComponent {
     const projects = this.drawerData().projectBreakdown ?? [];
     return formatNumber(projects.reduce((sum, p) => sum + p.sessions, 0));
   });
+
+  // === Paid media breakdown (from the revenueImpact input) ===
+  protected readonly paidMediaTrendChartData: Signal<ChartData<'bar'>> = this.initPaidMediaTrendChartData();
+  protected readonly projectBreakdownLegend: Signal<RevenueImpactChannelLegendView[]> = this.initProjectBreakdownLegend();
+  protected readonly sortedProjectBreakdown: Signal<RevenueImpactProjectBreakdownView[]> = this.initSortedProjectBreakdown();
+  protected readonly attributionChannelsView: Signal<RevenueImpactAttributionChannelView[]> = computed(() =>
+    this.data().attributionChannels.map((c) => ({
+      ...c,
+      label: PaidSocialReachDrawerComponent.formatChannelLabel(c.channel),
+      formattedPercentage: c.percentage.toFixed(1),
+    }))
+  );
 
   // === Protected Methods ===
   protected onClose(): void {
@@ -395,6 +488,106 @@ export class PaidSocialReachDrawerComponent {
           },
         ],
       };
+    });
+  }
+
+  private initProjectBreakdownLegend(): Signal<RevenueImpactChannelLegendView[]> {
+    return computed(() => {
+      const channelTotals = new Map<string, number>();
+      for (const r of this.data().projectBreakdown) {
+        for (const [channel, impressions] of Object.entries(r.channelImpressions)) {
+          channelTotals.set(channel, (channelTotals.get(channel) ?? 0) + (impressions ?? 0));
+        }
+      }
+      return Array.from(channelTotals.keys())
+        .sort((a, b) => (channelTotals.get(b) ?? 0) - (channelTotals.get(a) ?? 0))
+        .map((channel) => ({
+          channel,
+          label: PaidSocialReachDrawerComponent.formatChannelLabel(channel),
+          bgClass: PaidSocialReachDrawerComponent.bgClassFor(channel),
+        }));
+    });
+  }
+
+  private initSortedProjectBreakdown(): Signal<RevenueImpactProjectBreakdownView[]> {
+    return computed(() => {
+      const legend = this.projectBreakdownLegend();
+      return [...this.data().projectBreakdown]
+        .sort((a, b) => b.totalImpressions - a.totalImpressions)
+        .map((project) => {
+          const segments = legend
+            .map(({ channel, label, bgClass }) => {
+              const impressions = project.channelImpressions[channel] ?? 0;
+              const sharePercent = project.totalImpressions > 0 ? (impressions / project.totalImpressions) * 100 : 0;
+              return {
+                channel,
+                bgClass,
+                sharePercent,
+                title: `${label}: ${PaidSocialReachDrawerComponent.formatImpressionsShort(impressions)}`,
+              };
+            })
+            .filter((s) => s.sharePercent > 0);
+          return {
+            ...project,
+            formattedTotalImpressions: PaidSocialReachDrawerComponent.formatImpressionsShort(project.totalImpressions),
+            segments,
+          };
+        });
+    });
+  }
+
+  private initPaidMediaTrendChartData(): Signal<ChartData<'bar'>> {
+    return computed(() => {
+      const trend = this.data().paidMedia.monthlyTrend;
+      const labels = trend.map((r) => PaidSocialReachDrawerComponent.formatYearMonthLabel(r.month));
+      return {
+        labels,
+        datasets: [
+          {
+            type: 'bar',
+            label: 'Spend',
+            data: trend.map((r) => r.spend),
+            backgroundColor: lfxColors.blue[500],
+            borderRadius: 4,
+          },
+          {
+            type: 'bar',
+            label: 'Revenue',
+            data: trend.map((r) => r.revenue),
+            backgroundColor: lfxColors.emerald[600],
+            borderRadius: 4,
+          },
+        ],
+      };
+    });
+  }
+
+  private static formatChannelLabel(channel: string): string {
+    return channel
+      .split('_')
+      .map((word) => (word === 'ads' ? 'Ads' : word.charAt(0).toUpperCase() + word.slice(1)))
+      .join(' ');
+  }
+
+  private static formatImpressionsShort(n: number): string {
+    if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
+    if (n >= 1_000) return `${(n / 1_000).toFixed(0)}K`;
+    return n.toLocaleString();
+  }
+
+  private static bgClassFor(channel: string): string {
+    return PaidSocialReachDrawerComponent.channelBgClass[channel] ?? PaidSocialReachDrawerComponent.channelBgFallback;
+  }
+
+  private static formatYearMonthLabel(yearMonth: string): string {
+    const match = /^(\d{4})-(\d{2})$/.exec(yearMonth);
+    if (!match) return yearMonth;
+    const year = Number(match[1]);
+    const month = Number(match[2]) - 1;
+    return new Date(Date.UTC(year, month, 1)).toLocaleDateString('en-US', {
+      month: 'short',
+      year: '2-digit',
+      timeZone: 'UTC',
     });
   }
 }
