+++ apps/lfx-one/src/app/modules/dashboards/executive-director/components/revenue-impact-drawer/revenue-impact-drawer.component.html
@@ -261,94 +261,5 @@
     }
 
     <!-- Paid Media by Project -->
-    @if (data().projectBreakdown.length > 0) {
-      <div class="flex flex-col gap-4 p-4 border border-gray-200 rounded-lg" data-testid="revenue-impact-drawer-project-breakdown">
-        <div class="flex flex-col gap-1">
-          <h3 class="text-sm font-semibold text-gray-900">Paid Media by Project</h3>
-          <p class="text-sm text-gray-600">Impressions by project and channel over the last 6 months</p>
-        </div>
-
-        <div class="flex flex-wrap items-center gap-x-4 gap-y-2" data-testid="revenue-impact-drawer-project-breakdown-legend">
-          @for (legend of projectBreakdownLegend(); track legend.channel) {
-            <div class="flex items-center gap-2 text-xs text-gray-600">
-              <span class="inline-block w-3 h-3 rounded-sm" [class]="legend.bgClass"></span>
-              <span>{{ legend.label }}</span>
-            </div>
-          }
-        </div>
-
-        <div class="flex flex-col gap-0">
-          <div class="flex items-center justify-between py-2 px-1 text-xs font-semibold text-gray-500 uppercase border-b border-gray-200">
-            <span class="flex-1">Project</span>
-            <span class="flex-[2] px-2">Channel Mix</span>
-            <span class="w-32 text-right">Total Impressions</span>
-          </div>
-          @for (project of sortedProjectBreakdown(); track project.project) {
-            <div class="flex items-center justify-between py-3 px-1 border-b border-gray-100" data-testid="revenue-impact-drawer-project-breakdown-row">
-              <span class="flex-1 text-sm font-medium text-gray-900 truncate pr-2">{{ project.project }}</span>
-              <div class="flex-[2] px-2">
-                <div class="flex w-full h-3 rounded-sm overflow-hidden bg-gray-100">
-                  @for (segment of project.segments; track segment.channel) {
-                    <div class="h-full" [class]="segment.bgClass" [style.width.%]="segment.sharePercent" [title]="segment.title"></div>
-                  }
-                </div>
-              </div>
-              <span class="w-32 text-right text-sm font-semibold text-gray-900">{{ project.formattedTotalImpressions }}</span>
-            </div>
-          }
-        </div>
-
-        <p class="text-xs text-gray-400">Spend and revenue are only available at the foundation level, not per project.</p>
-      </div>
-    }
-
-    <!-- Attribution Channels -->
-    @if (attributionChannelsView().length > 0) {
-      <div class="flex flex-col gap-4 p-4 border border-gray-200 rounded-lg" data-testid="revenue-impact-drawer-channels">
-        <div class="flex flex-col gap-1">
-          <h3 class="text-sm font-semibold text-gray-900">Attribution Channels</h3>
-          <p class="text-sm text-gray-600">Impression share by paid media channel over the last 6 months</p>
-        </div>
-        <div class="flex flex-col gap-3">
-          @for (channel of attributionChannelsView(); track channel.channel) {
-            <div class="flex flex-col gap-1" data-testid="revenue-impact-drawer-channel-row">
-              <div class="flex items-center justify-between">
-                <span class="text-sm font-medium text-gray-900">{{ channel.label }}</span>
-                <span class="text-sm font-semibold text-gray-900">{{ channel.impressions | number }} · {{ channel.formattedPercentage }}%</span>
-              </div>
-              <div class="w-full h-2 bg-gray-100 rounded-full">
-                <div class="h-2 bg-blue-500 rounded-full" [style.width.%]="channel.percentage"></div>
-              </div>
-            </div>
-          }
-        </div>
-      </div>
-    }
-
-    <!-- SECTION 2: Paid Media (Marketing Attribution) -->
-    @if (data().paidMedia.monthlyTrend.length > 1) {
-      <div class="flex flex-col gap-4" data-testid="revenue-impact-drawer-paid-media">
-        <h3 class="text-sm font-semibold text-gray-900">Paid Media</h3>
-        <div class="flex flex-col gap-4 p-4 border border-gray-200 rounded-lg" data-testid="revenue-impact-drawer-paid-media-trend">
-          <div class="flex flex-col gap-1">
-            <h3 class="text-sm font-semibold text-gray-900">Paid Media Trend</h3>
-            <p class="text-sm text-gray-600">Monthly spend vs first-touch revenue over the last 6 months</p>
-          </div>
-          <div class="h-[240px]" data-testid="revenue-impact-drawer-paid-media-chart">
-            <lfx-chart type="bar" [data]="paidMediaTrendChartData()" [options]="paidMediaTrendChartOptions" height="100%"></lfx-chart>
-          </div>
-          <div class="grid grid-cols-2 gap-2 text-xs text-gray-500">
-            <div class="flex items-center gap-2">
-              <span class="inline-block w-3 h-3 rounded-sm bg-blue-500"></span>
-              <span>Spend — dollars per month</span>
-            </div>
-            <div class="flex items-center gap-2">
-              <span class="inline-block w-3 h-3 rounded-sm bg-emerald-600"></span>
-              <span>Revenue — first-touch attributed revenue per month</span>
-            </div>
-          </div>
-        </div>
-      </div>
-    }
   </div>
 </p-drawer>
